changes on the structure of the class Loan
Adds calcPenalty: calculates de amount of money to be payed based on the delay on the return of the item and it's original cost. Adds isLate function: verifies if the item is late or not, if it is it deducts the penalty from the clients balance. Adds constructor method. Still a scratch, methods not complete nor checked.