Changes in the Client class and Loan methods
Adds hasLateItems to check if the client has any late items unreturned. Adds rentItem: the client can only rent an item if there are items left to rent, if the client is not in debt with the Locator and if he hasn't reached the limit of items simultaneously borrowed. Altered the Loans' method returnItem: when the client returns the item, this loan is removed from the Client's ArrayLists of simultaneous loans.