updates renew function
Adds private calcRenewCost method: calculates the cost to renew a loan based on how many days are going to be added to the deadline. Updated renewItem method, to call the calcRenewCost method and update deadline and total cost. Added an atribute renewed(boolean): check if the loan has been renewed yet or not.