changes in Menu -> menuClient
Added a menu client method for the user to be able to be able to sign up, sign in and loan a product. Only implemented sign up method so far. Very similar to addClient so it is a little reduntant, verify and correct it. Needs a library ValidatesDocuments, to validate CPF. In RentalStore, I implemented a method alreadySigned to verify if the client has already signed up to the store, by searching for their cpf in the clients ArrayList